requirements.txt
colorama==0.4.4
requests==2.26.0
dnspython==2.2.1
scapy==2.4.5
beautifulsoup4==4.10.0
python-nmap==0.7.1
tqdm==4.62.3